Make request to open weather API service, and fill the fields in CityDetailActiviry according to the response

+++ app/src/main/java/com/poviolab/poviolabtest/model/CityLab.java
@@ -33,7 +33,7 @@ public class CityLab {
         return mCities;
     }
 
-    public City getCity(int i) {
-        return mCities.get(i);
+    public City getCity(long i) {
+        return mCities.get((int)i);
     }
 }
